+++ electrum/lntransport.py
@@ -5,19 +5,23 @@
 
 # Derived from https://gist.github.com/AdamISZ/046d05c156aaeb56cc897f85eecb3eb8
 
+import re
 import hashlib
 import asyncio
 from asyncio import StreamReader, StreamWriter
 from typing import Optional
 from functools import cached_property
+from typing import NamedTuple, List, Tuple, Mapping, Optional, TYPE_CHECKING, Union, Dict, Set, Sequence
 
+from aiorpcx import NetAddress
 import electrum_ecc as ecc
 
-from .crypto import sha256, hmac_oneshot, chacha20_poly1305_encrypt, chacha20_poly1305_decrypt
-from .lnutil import (get_ecdh, privkey_to_pubkey, LightningPeerConnectionClosed,
-                     HandshakeFailed, LNPeerAddr)
+from .crypto import sha256, hmac_oneshot, chacha20_poly1305_encrypt, chacha20_poly1305_decrypt, get_ecdh, privkey_to_pubkey
 from .util import MySocksProxy
 
+class LightningPeerConnectionClosed(Exception): pass
+class HandshakeFailed(Exception): pass
+class ConnStringFormatError(Exception): pass
 
 class HandshakeState(object):
     prologue = b"lightning"
@@ -89,6 +93,99 @@ def create_ephemeral_key() -> (bytes, bytes):
     privkey = ecc.ECPrivkey.generate_random_key()
     return privkey.get_secret_bytes(), privkey.get_public_key_bytes()
 
+
+def split_host_port(host_port: str) -> Tuple[str, str]: # port returned as string
+    ipv6  = re.compile(r'\[(?P<host>[:0-9a-f]+)\](?P<port>:\d+)?$')
+    other = re.compile(r'(?P<host>[^:]+)(?P<port>:\d+)?$')
+    m = ipv6.match(host_port)
+    if not m:
+        m = other.match(host_port)
+    if not m:
+        raise ConnStringFormatError('Connection strings must be in <node_pubkey>@<host>:<port> format')
+    host = m.group('host')
+    if m.group('port'):
+        port = m.group('port')[1:]
+    else:
+        port = '9735'
+    try:
+        int(port)
+    except ValueError:
+        raise ConnStringFormatError('Port number must be decimal')
+    return host, port
+
+def extract_nodeid(connect_contents: str) -> Tuple[bytes, Optional[str]]:
+    """Takes a connection-string-like str, and returns a tuple (node_id, rest),
+    where rest is typically a host (with maybe port). Examples:
+    - extract_nodeid(pubkey@host:port) == (pubkey, host:port)
+    - extract_nodeid(pubkey@host) == (pubkey, host)
+    - extract_nodeid(pubkey) == (pubkey, None)
+    Can raise ConnStringFormatError.
+    """
+    rest = None
+    try:
+        # connection string?
+        nodeid_hex, rest = connect_contents.split("@", 1)
+    except ValueError:
+        # node id as hex?
+        nodeid_hex = connect_contents
+    if rest == '':
+        raise ConnStringFormatError('At least a hostname must be supplied after the at symbol.')
+    try:
+        node_id = bytes.fromhex(nodeid_hex)
+        if len(node_id) != 33:
+            raise Exception()
+    except Exception:
+        raise ConnStringFormatError('Invalid node ID, must be 33 bytes and hexadecimal')
+    return node_id, rest
+
+class LNPeerAddr:
+    # note: while not programmatically enforced, this class is meant to be *immutable*
+
+    def __init__(self, host: str, port: int, pubkey: bytes):
+        assert isinstance(host, str), repr(host)
+        assert isinstance(port, int), repr(port)
+        assert isinstance(pubkey, bytes), repr(pubkey)
+        try:
+            net_addr = NetAddress(host, port)  # this validates host and port
+        except Exception as e:
+            raise ValueError(f"cannot construct LNPeerAddr: invalid host or port (host={host}, port={port})") from e
+        # note: not validating pubkey as it would be too expensive:
+        # if not ECPubkey.is_pubkey_bytes(pubkey): raise ValueError()
+        self.host = host
+        self.port = port
+        self.pubkey = pubkey
+        self._net_addr = net_addr
+
+    def __str__(self):
+        return '{}@{}'.format(self.pubkey.hex(), self.net_addr_str())
+
+    @classmethod
+    def from_str(cls, s):
+        node_id, rest = extract_nodeid(s)
+        host, port = split_host_port(rest)
+        return LNPeerAddr(host, int(port), node_id)
+
+    def __repr__(self):
+        return f'<LNPeerAddr host={self.host} port={self.port} pubkey={self.pubkey.hex()}>'
+
+    def net_addr(self) -> NetAddress:
+        return self._net_addr
+
+    def net_addr_str(self) -> str:
+        return str(self._net_addr)
+
+    def __eq__(self, other):
+        if not isinstance(other, LNPeerAddr):
+            return False
+        return (self.host == other.host
+                and self.port == other.port
+                and self.pubkey == other.pubkey)
+
+    def __ne__(self, other):
+        return not (self == other)
+
+    def __hash__(self):
+        return hash((self.host, self.port, self.pubkey))
 
 class LNTransportBase:
     reader: StreamReader
